dev → main 프로덕션 배포

+++ user-service/src/main/java/com/sajo/user_service/account/client/kis/dto/response/KisAccessTokenResponse.java
@@ -3,7 +3,7 @@ package com.sajo.user_service.account.client.kis.dto.response;
 public record KisAccessTokenResponse(
         String access_token,
         String token_type,
-        float expires_in,
+        long expires_in,
         String access_token_token_expired
 
 ){
